+++ lib/avromatic/model/passthrough_serializer.rb
@@ -1,0 +1,10 @@
+module Avromatic
+  module Model
+    # This trivial serializer simply returns the value provided.
+    module PassthroughSerializer
+      def self.call(value)
+        value
+      end
+    end
+  end
+end