+++ spec/avro/dsl/test/with_varchar.rb
@@ -1,0 +1,10 @@
+namespace :test
+
+record :varchar do
+  required :length, :int
+  required :data, :bytes
+end
+
+record :with_varchar do
+  required :str, :varchar
+end