+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # avromatic changelog
 
+## v0.4.0
+- Allow the specification of a custom type, including conversion to/from Avro,
+  for named types.
+
 ## v0.3.0
 - Remove dependency on the `private_attr` gem.